+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Deprecate the option `debug`. `verbose` should be used instead.
 * Change YUI option names to clearly indicate that they are specific to YUI Compressor.
 * Update Google Closure Compiler to v20130823.
+* Add support for Google Closure Compiler `language` option.
 
 ## 1.7.1